Module aws_sdk_sagemakerfeaturestoreruntime::error

Modules

access_forbidden

See AccessForbidden

internal_failure

See InternalFailure

resource_not_found

See ResourceNotFound

service_unavailable

See ServiceUnavailable

validation_error

See ValidationError

Structs

AccessForbidden

You do not have permission to perform an action.

BatchGetRecordError

DeleteRecordError

GetRecordError

InternalFailure

An internal failure occurred. Try your request again. If the problem

PutRecordError

ResourceNotFound

A resource that is required to perform an action was not found.

ServiceUnavailable

The service is currently unavailable.

ValidationError

There was an error validating your request.

Enums

BatchGetRecordErrorKind

DeleteRecordErrorKind

GetRecordErrorKind

PutRecordErrorKind